@@ -1,9 +1,15 @@
-<Project Sdk="Microsoft.NET.Sdk">
+﻿<Project Sdk="Microsoft.NET.Sdk">
 
   <PropertyGroup>
     <OutputType>Exe</OutputType>
-    <TargetFramework>netcoreapp3.0</TargetFramework>
+    <TargetFramework>netcoreapp3.1</TargetFramework>
   </PropertyGroup>
+
+  <ItemGroup>
+    <Compile Remove="bin\Release\netcoreapp3.1\tests\**" />
+    <EmbeddedResource Remove="bin\Release\netcoreapp3.1\tests\**" />
+    <None Remove="bin\Release\netcoreapp3.1\tests\**" />
+  </ItemGroup>
 
   <ItemGroup>
     <ProjectReference Include="..\NESEmulator\NESEmulator.csproj" />
@@ -17,14 +23,10 @@
     </None>
   </ItemGroup>
   <ItemGroup>
-    <PackageReference Include="OpenTK.NetStandard" Version="1.0.4" />
+    <PackageReference Include="OpenTK.NetStandard" Version="1.0.5.12" />
   </ItemGroup>
   <ItemGroup>
+    <Folder Include="bin\Release\netcoreapp3.1\" />
     <Folder Include="lib\" />
-  </ItemGroup>
-  <ItemGroup>
-    <Reference Include="csPixelGameEngine">
-      <HintPath>lib\csPixelGameEngine.dll</HintPath>
-    </Reference>
   </ItemGroup>
 </Project>
